# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/locations/countries/{country_id}": {
            "get": {
                "description": "GetCoundry by countryID (external api)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetCountry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country ID",
                        "name": "country_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/locations.Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ApiError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "locations.Country": {
            "type": "object",
            "properties": {
                "geo_information": {
                    "$ref": "#/definitions/locations.GeoInformation"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "states": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/locations.State"
                    }
                },
                "time_zone": {
                    "type": "string"
                }
            }
        },
        "locations.GeoInformation": {
            "type": "object",
            "properties": {
                "location": {
                    "$ref": "#/definitions/locations.GeoLocation"
                }
            }
        },
        "locations.GeoLocation": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "locations.State": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}